Рефакторинг переливания, интеграция внутрь бота

+++ fucking_game/handlers/fill_undef_values.py
@@ -3,7 +3,8 @@ from aiogram.types import Message, CallbackQuery, BufferedInputFile
 from aiogram.fsm.context import FSMContext
 from aiogram.exceptions import TelegramBadRequest
 
-from flasks import flasks_solver
+# from flasks import flasks_solver
+from transfusion_of_liquids import transfusion_manage
 from found_colors import found_colors_in_flasks, replace_in_json, create_image_for_replace, add_empty_flask
 import config
 import classes.all_my_classes as amc
@@ -140,13 +141,15 @@ async def fill_undef_values(callback: CallbackQuery, state: FSMContext):
 
         logger.log_info(f'Пользователь {callback.from_user.id} заполнил все пустоты')
         try:
+            # Вызываем функцию перебора переливаний
+            is_solved = transfusion_manage(task=in_file, result=out_file)
             # Запускаем файл для решения уровня
-            flasks_solver(input_file=in_file, output_file=out_file)
+            # flasks_solver(input_file=in_file, output_file=out_file)
         except TelegramBadRequest:
             logger.log_error('Превышено время ожидания ответа на начало поиска решения')
 
         # В случае, если файл пустой или не был создан сообщаем, что решение не найдено, иначе выводим решение
-        if os.stat(out_file).st_size == 0 or os.path.isfile(out_file) == False:
+        if not is_solved:
             await callback.message.answer(
                 f'😖😖😖Unfortunately, I was unable to find a solution for this arrangement.\nIf you want to change the order of undefined colors, click "🔄️🖼️Reload image".\nIf you know all the colors, but the solution still hasn’t been found, then I can add another empty flask, to do this, click “➕🧪Add an empty flask”\nOr you can upload a new image, to do this, click "📩🖼️Upload new image"',
                 reply_markup=no_result()
@@ -160,7 +163,8 @@ async def fill_undef_values(callback: CallbackQuery, state: FSMContext):
                 )
             await state.set_state(amc.SolveFlasks.start_solving)
         # Удаление временных файлов
-        os.remove(out_file)
+        if os.path.isfile(out_file):
+            os.remove(out_file)
         os.remove(in_file)
         os.remove(lvl_file)
