+++ fucking_game/handlers/get_photo.py
@@ -3,7 +3,8 @@ from aiogram.types import Message, BufferedInputFile
 from aiogram.fsm.context import FSMContext
 from aiogram.exceptions import TelegramBadRequest
 
-from flasks import flasks_solver
+# from flasks import flasks_solver
+from transfusion_of_liquids import transfusion_manage
 from found_colors import found_colors_in_flasks, create_image_for_replace
 import classes.all_my_classes as amc
 from keyboards.all_my_keyboards import error_image, colors, feedback, upload_new, no_result
@@ -94,13 +95,15 @@ async def get_photo(message: Message, bot: Bot, state: FSMContext):
             )
         logger.log_info(f'Пользователь {message.from_user.id} заполнил все пустоты')
         try:
+            # Вызываем функцию перебора переливаний
+            is_solved = transfusion_manage(task=in_file, result=out_file)
             # Запускаем файл для решения уровня
-            flasks_solver(input_file=in_file, output_file=out_file)
+            # flasks_solver(input_file=in_file, output_file=out_file)
         except TelegramBadRequest:
             logger.log_error('Превышено время ожидания ответа на начало поиска решения')
 
         # В случае, если файл пустой или не был создан сообщаем, что решение не найдено, иначе выводим решение
-        if os.stat(out_file).st_size == 0 or os.path.isfile(out_file) == False:
+        if not is_solved:
             await message.answer(
                 f"😖😖😖Unfortunately, I was unable to find a solution for this arrangement.\nIf you want to change the order of undefined colors, click '🔄️🖼️Reload image'.\nIf you know all the colors, but the solution still hasn't been found, then I can add another empty flask, to do this, click '➕🧪Add an empty flask'\nOr you can upload a new image, to do this, click '📩🖼️Upload new image'",
                 reply_markup=no_result()
@@ -114,7 +117,8 @@ async def get_photo(message: Message, bot: Bot, state: FSMContext):
                 )
             await state.set_state(amc.SolveFlasks.start_solving)
         # Удаление временных файлов
-        os.remove(out_file)
+        if os.path.isfile(out_file):
+            os.remove(out_file)
         os.remove(in_file)
         os.remove(lvl_file)
